+++ insteon_mqtt/mqtt/IOLinc.py
@@ -38,7 +38,7 @@ class IOLinc:
             payload='{ "cmd" : "{{value.lower()}}" }')
 
         # Input scene on/off command template.
-        self.msg_scene_on_off = MsgTemplate(
+        self.msg_scene = MsgTemplate(
             topic='insteon/{{address}}/scene',
             payload='{ "cmd" : "{{value.lower()}}" }')
 
@@ -60,8 +60,7 @@ class IOLinc:
         self.msg_state.load_config(data, 'state_topic', 'state_payload', qos)
         self.msg_on_off.load_config(data, 'on_off_topic', 'on_off_payload',
                                     qos)
-        self.msg_scene_on_off.load_config(data, 'scene_on_off_topic',
-                                          'scene_on_off_payload', qos)
+        self.msg_scene.load_config(data, 'scene_topic', 'scene_payload', qos)
 
     #-----------------------------------------------------------------------
     def subscribe(self, link, qos):
@@ -79,7 +78,7 @@ class IOLinc:
         link.subscribe(topic, qos, self._input_on_off)
 
         # Scene triggering messages.
-        topic = self.msg_scene_on_off.render_topic(self.template_data())
+        topic = self.msg_scene.render_topic(self.template_data())
         link.subscribe(topic, qos, self._input_scene)
 
     #-----------------------------------------------------------------------
@@ -92,7 +91,7 @@ class IOLinc:
         topic = self.msg_on_off.render_topic(self.template_data())
         link.unsubscribe(topic)
 
-        topic = self.msg_scene_on_off.render_topic(self.template_data())
+        topic = self.msg_scene.render_topic(self.template_data())
         link.unsubscribe(topic)
 
     #-----------------------------------------------------------------------
@@ -177,7 +176,7 @@ class IOLinc:
         LOG.debug("IOLinc message %s %s", message.topic, message.payload)
 
         # Parse the input MQTT message.
-        data = self.msg_scene_on_off.to_json(message.payload)
+        data = self.msg_scene.to_json(message.payload)
         LOG.info("IOLinc input command: %s", data)
 
         try: