+++ tests/mqtt/test_IOLinc.py
@@ -155,8 +155,8 @@ class Test_IOLinc:
 
         qos = 2
         config = {'io_linc' : {
-            'scene_on_off_topic' : 'foo/{{address}}',
-            'scene_on_off_payload' : '{ "cmd" : "{{json.on.lower()}}" }'}}
+            'scene_topic' : 'foo/{{address}}',
+            'scene_payload' : '{ "cmd" : "{{json.on.lower()}}" }'}}
         mdev.load_config(config, qos=qos)
 
         mdev.subscribe(link, qos)