+++ HISTORY.md
@@ -7,6 +7,10 @@
 
 - Fixed incorrect exception log statement in Protocol ([Issue #132][I132])
 
+- Fixed incorrect scene names in the config loader for IOLinc, Outlet, and
+  Switch.  This prevented customizing the MQTT scene topic and payload for
+  those devices ([Issue #130][I130])
+
 
 ## [0.6.7]
 
@@ -254,4 +258,5 @@
 [I119]: https://github.com/TD22057/insteon-mqtt/issues/119
 [I120]: https://github.com/TD22057/insteon-mqtt/issues/120
 [I126]: https://github.com/TD22057/insteon-mqtt/issues/126
+[I130]: https://github.com/TD22057/insteon-mqtt/issues/130
 [I132]: https://github.com/TD22057/insteon-mqtt/issues/132